+++ src/Cns.test.ts
@@ -173,4 +173,14 @@ describe('CNS', () => {
       expect(new Resolution().namehash('crypto')).toEqual('0x0f4a10a4f46c288cea365fcf45cccf0e9d901b945b9829ccdb54c10dc3cb7a6f');
     });
   });
+
+  describe('meta data', () => {
+    it('should resolve with correct metaData', async () => {
+      const resolution = new Resolution();
+      const ipfs = await resolution.ipfs(labelDomain);
+      const whois = await resolution.whois(labelDomain);
+      console.log({ipfs, whois});
+
+    })
+  })
 });