+++ src/Zns.test.ts
@@ -5,6 +5,7 @@ import {
   ZilliqaUrl,
   mockAPICalls,
 } from './utils/testHelpers';
+import ResolutionError from './resolutionError';
 
 beforeEach(() => {
   jest.restoreAllMocks();
@@ -286,4 +287,21 @@ describe('ZNS', () => {
   //     },
   //   });
   // });
+
+  describe('Meta data', () => {
+    it('should return whois method', async () => {
+      const resolution = new Resolution();
+      const whois = await resolution.whois('ergergergerg.zil');
+      expect(whois).toStrictEqual({ email: 'matt+test@unstoppabledomains.com',
+      url: 'www.unstoppabledomains.com',
+      for_sale: true });
+    });
+
+    it('should return ipfs method', async () => {
+      const resolution = new Resolution();
+      const ipfs = await resolution.ipfs('ergergergerg.zil');
+      expect(ipfs).toStrictEqual({ hash: 'QmVaAtQbi3EtsfpKoLzALm6vXphdi2KjMgxEDKeGg6wHu',
+      redirect: 'www.unstoppabledomains.com' } );
+    });
+  });
 });